refactor: improve integer types and conversions + various minor improvements (#337)
This refactoring reduces the use of `usize` in the encoding code and
clarifies our assumptions on the underlying architecture. This was
triggered by this discussion:
https://github.com/MystenLabs/walrus/pull/329#discussion_r1583245447.

This also replaces conversion between integer types using `as` by
`from`, `try_from`, `into`, and `try_into` to clearly state why we know
or expect the conversion to succeed and be lossless.

The PR also includes a few other minor improvements:
- Relax trait bounds on `walrus_test_utils::random_subset*`.
- Improve some error types.
- Add/improve some docstrings.
- Change metadata input to `Sliver::verify`.
- Add a few convenience functions.

+++ crates/walrus-core/src/encoding/common.rs
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 /// The maximum length in bytes of a single symbol in RaptorQ.
-pub const MAX_SYMBOL_SIZE: usize = u16::MAX as usize;
+pub const MAX_SYMBOL_SIZE: u16 = u16::MAX;
 
 /// The maximum number of source symbols per block for RaptorQ.
 ///
